add mapping for channels (nwp and satellite)
