+++ nowcasting_dataloader/data_sources/satellite/satellite_model.py
@@ -9,7 +9,10 @@ from nowcasting_dataset.time import make_random_time_vectors
 from pydantic import Field
 
 from nowcasting_dataloader.data_sources.datasource_output import Array, DataSourceOutputML
-from nowcasting_dataloader.utils.dims import re_order_dims
+from nowcasting_dataloader.xr_utils import re_order_dims, map_channels_names_to_indexes
+from nowcasting_dataset.consts import SAT_VARIABLE_NAMES
+
+channels_mapping = {key: i for i, key in enumerate(SAT_VARIABLE_NAMES)}
 
 logger = logging.getLogger(__name__)
 
@@ -113,6 +116,9 @@ class SatelliteML(DataSourceOutputML):
     @staticmethod
     def from_xr_dataset(xr_dataset: xr.Dataset):
         """Change xr dataset to model."""
+        
+        # change channel names to indexes
+        xr_dataset = map_channels_names_to_indexes(xr_dataset, channels_mapping)
 
         # make sure the dims are in the correct order
         xr_dataset = re_order_dims(xr_dataset)
